Access to all of PGP and types
Added getMain() where you have access to pgp so you can use utilities, helpers, etc. 
getMain() works the same way as getPg() does, using the initial configuration to create it only the first time
getPg() was changed to instead of instanciating pgp, it delegates the task to getMain() and then uses the instance.